Migrate Filter Status into Own Trait
